# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/looking-for": {
      "post": {
        "operationId": "LookingForController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLookingForDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Looking for category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/LookingFor"
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Create a new looking for category",
        "tags": [
          "looking-for"
        ]
      },
      "get": {
        "operationId": "LookingForController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of items to take",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by name",
            "schema": {
              "example": "web",
              "type": "string"
            }
          },
          {
            "name": "industrySlug",
            "required": false,
            "in": "query",
            "description": "Industry slug to filter by",
            "schema": {
              "example": "technology",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of looking for categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookingFor"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all looking for categories with pagination and filtering",
        "tags": [
          "looking-for"
        ]
      }
    },
...
    "/looking-for/{id}": {
      "get": {
        "operationId": "LookingForController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Looking for category ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Looking for category retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookingFor"
                }
              }
            }
          },
          "404": {
            "description": "Looking for category not found"
          }
        },
        "summary": "Get a looking for category by ID",
        "tags": [
          "looking-for"
        ]
      },
      "patch": {
        "operationId": "LookingForController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Looking for category ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLookingForDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Looking for category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookingFor"
                }
              }
            }
          },
          "404": {
            "description": "Looking for category not found"
          }
        },
        "summary": "Update a looking for category",
        "tags": [
          "looking-for"
        ]
      },
      "delete": {
        "operationId": "LookingForController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Looking for category ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Looking for category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookingFor"
                }
              }
            }
          },
          "404": {
            "description": "Looking for category not found"
          }
        },
        "summary": "Delete a looking for category",
        "tags": [
          "looking-for"
        ]
      }
    }
  },
  "info": {
    "title": "DEPA API",
    "description": "The DEPA API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLookingForDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLookingForDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}